requirements.txt
pyYAML==5.3.1
requests==2.25.0